Refactor: view (inputView) 분리
