+++ src/utils/LottoPurchased.js
@@ -1,16 +1,15 @@
 import { Random, Console } from "@woowacourse/mission-utils";
 import { CONSOLE_MESSAGE, ERROR_MESSAGE } from "../Constants.js";
 import Purchase from "../errors/Purchase.js";
+import { InputView } from "../view/inputVeiw.js";
 class LottoPurchased {
   async getPurchaseQuantity() {
     let purchaseAmount;
 
     while (true) {
       try {
-        purchaseAmount = await Console.readLineAsync(
-          CONSOLE_MESSAGE.PURCHASE_AMOUNT_INPUT + "\n"
-        );
-        new Purchase(purchaseAmount);
+        purchaseAmount = await InputView.readPurchaseAmount();
+
         break;
       } catch (error) {
         Console.print(ERROR_MESSAGE.NOT_DIVISIBLE_BY_1000);
@@ -25,7 +24,7 @@ class LottoPurchased {
   async getLottoNumbers(quantity, min, max, count) {
     let myLottos = [];
 
-    for (let count = 0; count < quantity; count++) {
+    for (let i = 0; i < quantity; i++) {
       const myLotto = await Random.pickUniqueNumbersInRange(min, max, count);
       myLotto.sort((a, b) => a - b);
       myLottos.push(myLotto);
